+++ docs/src/examples/components/Popup/Rtl/PopupExample.rtl.tsx
@@ -1,15 +1,18 @@
 import * as React from 'react'
-import { Button, Popup } from '@stardust-ui/react'
+import { Button, Popup, Text } from '@stardust-ui/react'
 
 const PopupExampleRtl = () => (
   <Popup
     trigger={<Button content="ا يجلبه إلينا الأس" />}
     content={{
       content: (
-        <p>
-          ينا الألم. في بعض الأحيان ونظراً للالتزامات التي يفرضها علينا الواجب والعمل سنتنازل غالباً
-          ونرفض الشعور
-        </p>
+        <>
+          <Text as="div">English text!</Text>
+          <Text as="div">
+            ينا الألم. في بعض الأحيان ونظراً للالتزامات التي يفرضها علينا الواجب والعمل سنتنازل
+            غالباً ونرفض الشعور
+          </Text>
+        </>
       ),
     }}
   />