+++ packages/react/src/themes/base/components/Provider/providerBoxStyles.ts
@@ -4,5 +4,6 @@ export default {
   root: ({ variables: v }): ICSSInJSStyle => ({
     background: v.background,
     color: v.color,
+    textAlign: 'left',
   }),
 }