+++ packages/react/src/components/Ref/RefForward.tsx
@@ -1,7 +1,8 @@
+import * as customPropTypes from '@stardust-ui/react-proptypes'
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
-import { ChildrenComponentProps, customPropTypes, handleRef } from '../../lib'
+import { ChildrenComponentProps, handleRef } from '../../lib'
 
 export interface RefForwardProps extends ChildrenComponentProps<React.ReactElement<any>> {
   /**